Allow multiple branches between buses; remove objects; import SciGRID example
Multiple branches between buses are allowed via nx.MultiGraph; the
branch object is the key for the edge.

Objects can now be gracefully removed.

A network.topology_determined switch indicates whether the topology
has already been calculated / needs to be recalculated.

An example of importing SciGRID is included.

Changes to be committed:
	modified:   .gitignore
	modified:   components.py
	modified:   descriptors.py
	new file:   examples/scigrid-151109/LICENSE
	new file:   examples/scigrid-151109/NOTICE
	new file:   examples/scigrid-151109/import_scigrid.py
	new file:   examples/scigrid-151109/links_de_power_151109.csvdata
	new file:   examples/scigrid-151109/vertices_de_power_151109.csvdata
	new file:   examples/scigrid-151109/vertices_ref_id.csvdata
	new file:   guidelines.org
	modified:   io.py
	modified:   opf.py
	modified:   pf.py