Add first version of Store component
Store is more primitive than StorageUnit and can be used to move
energy between snapshots. It can be combined with a lossly link to
make a generator; it can be combined with two links to make a storage
unit.

So far it has only been added to components.py and opf.py.