opf: Fix strange bug in reindexing of dual keys
For some reason, unless the index passed to a pandas Series is wrapped
in pd.Index, it maps all pyomo SimpleConstraints (i.e. constraints
which are NOT indexed, like the co2_constraint) to (None,), making the
index non-unique if there is more than one SimpleConstraint and
causing problems with the .map reindex.