Workaround for a bad interaction between pandas and Pyomo 4.4.1
Due to new code in Pyomo 4.4.1, the first index set passed to Pyomo
Var or Constraint cannot be a pandas.Index - it must be converted to a
list.

In

pyomo/core/base/var.py

there is a check in Var.__new__:

(args[0] == UnindexedComponent_set and len(args)==1)

If args[0] is a pandas.Index, then the first comparison will be a
Boolean np.array, which cannot be combined with a single Boolean.