components, io: Avoid duplicates of standard types
Because PyPSA now has default standard types, we must avoid
duplication when reading them in, exporting them or copying
them. Otherwise this leads to tedious errors.

The default behaviour is to always read them in when the network is
initialised and to drop them again when the network is exported, so
that only user-defined types are exported.