components, opf: Add new global constraint component; remove co2 limit
This generalises the co2 limit to more general global constraints on
resource usage. This constraint can be on any property of a carrier
and can have any sense <=, == or >=.

The shadow price is updated and stored in
network.global_constraints.mu.

The examples ac-dc-meshed and opf-storage-hvdc have been updated
appropriately.