linopf: Fix bug whereby snapshot_weightings name is missing
The old code relied on snapshot_weightings.name being
"weightings". However, this is only the case when it has been exported
to a netcdf file (when it's given this name) and reimported.

Networks created without exporting and importing break at this point
with emissions constraints. Minimum breaking model is provided in the
pull request.