linopf.py: Fix bug: multi-link outputs set when not multi-link
There can be mixed situations where some links have multiple outputs
to bus2, bus3, etc, but we still have some links with single outputs
to bus1 and no other outputs bus2 = "" and bus3 = "".

However, the LOPF with pyomo=False was setting non-zero outputs for p2
and p3 also for the links with single outputs.

Now, if bus2 and bus3 etc are empty, the p2 and p3 is set to the
default, just like pyomo=True.