Exclude store links from fixing p_nom at last iteration
The last linopf iteration fixes the link and line capacities to ensure consistency of impedances and power flow. However, when using store components, it also fixes the capacities of their respective in- and outflow links. Filtering by carrier solves this problem.
This is a quick fix which has worked well for me, I'm sure there's a more elegant solution than using an if loop - however, it is nice that no error messages are displayed, using the loc approach.