Update linopf.py
Sorry for the delay, I just now figured out how to re-commit .. the indents have been added, thank you!

Moreover, I cautiously added the before mentioned approach to include the capital costs of link + line expansion to the objective value of the output network. They have not yet been included, since they are not considered as expansion during the last iteration. Please feel free to verify yourselves that this has not been reflected in the n.objective of the output networks before. (Furthermore, it should be noted for verification that if using store components, the costs for expansion of (dis)charge links have not been reflected either since their capacity was also fixed + non-extendable. However, this is solved with the initial commit of this pull request, thus, it is filtered by DC again to avoid duplication of costs). Please feel free to correct me if this is not right, but at least for my PyPSA-EUR output networks, this is definitely the case.