Exclude store links from fixing p_nom at last iteration (#243)
* Exclude store links from fixing p_nom at last iteration

The last linopf iteration fixes the link and line capacities to ensure consistency of impedances and power flow. However, when using store components, it also fixes the capacities of their respective in- and outflow links. Filtering by carrier solves this problem.
This is a quick fix which has worked well for me, I'm sure there's a more elegant solution than using an if loop - however, it is nice that no error messages are displayed, using the loc approach.

* Update pypsa/linopf.py

Co-authored-by: Fabian Neumann <fabian.neumann@outlook.de>

* Update linopf.py

Sorry for the delay, I just now figured out how to re-commit .. the indents have been added, thank you!

Moreover, I cautiously added the before mentioned approach to include the capital costs of link + line expansion to the objective value of the output network. They have not yet been included, since they are not considered as expansion during the last iteration. Please feel free to verify yourselves that this has not been reflected in the n.objective of the output networks before. (Furthermore, it should be noted for verification that if using store components, the costs for expansion of (dis)charge links have not been reflected either since their capacity was also fixed + non-extendable. However, this is solved with the initial commit of this pull request, thus, it is filtered by DC again to avoid duplication of costs). Please feel free to correct me if this is not right, but at least for my PyPSA-EUR output networks, this is definitely the case.

* Update linopf.py

Excluding all links from being fixed + adjusting the post-processing of n.objective for the additional AC infrastructure

* Revert to earlier version where DC links are fixed

+ remove what is added to objective from objective_constant

Co-authored-by: @chiaroo

* minor code optimisation

* Bugfix: copy s_nom_opt and s_nom for s_nom_prev

* Update release_notes.rst

Co-authored-by: Fabian Neumann <fabian.neumann@outlook.de>