Fix some line parameter calculation issues in network clustering
Update the network components attributes for lines so that r, x will not be missing in aggregatelines function.
As length is an optional input when type is not set, populate the orginal network with length information if it is not there so that length factors can be properly cauculated
Testing for input branches with 0 length but non-zero paramters, which could exist under certain conditions. Also make it forgiving to avoid 0-dividing when calculating parameters aggregatelines.
Carry forward global constraints after clustering.