Update constraints.py
In the nodal balance constraint there is a check if the nodal balance is empty and the rhs!=0. For optimisations with snapshots as a multi-index and an empty nodal balance this currently breaks with the error message
ValueError: Unable to coerce to DataFrame
This is due to the fact that the rhs and the empty nodal balance do not align. E.g. empty _nodal_balance
```
<xarray.DataArray 'vars' (Bus: 364, snapshot: 4)>
array([[ True, False, False, False],
       [False, False, False, False],
       [False, False, False, False],
       ...,
       [False, False, False, False],
       [False, False, False, False],
       [False, False, False, False]])
Coordinates:
  * Bus       (Bus) object 'AL1 0 EV battery' ... 'solid biomass for industry'
  * snapshot  (snapshot) object MultiIndex
  * period    (snapshot) int64 2020 2030 2040 2050
  * timestep  (snapshot) datetime64[ns] 2013-01-01 2013-01-01 ... 2013-01-01
```
rhs
```
Bus                AL1 0 EV battery  ...  solid biomass for industry
period timestep                      ...                            
2020   2013-01-01          0.000000  ...                80182.648402
2030   2013-01-01         56.678867  ...                80182.648402
2040   2013-01-01        136.029280  ...                80182.648402
2050   2013-01-01        192.708146  ...                80182.648402

[4 rows x 364 columns]

```
This can be fixed be transposing the empty_nodal_balance as suggested. @FabianHofmann what do you think?