new tests and suggestions integrated
Updated method to include additional inputs suggested by Fabian and error thrown if snapshots/snapshot weightings in the networks do not agree. Also added warning if components with IDs matching component in existing network are added - this warning does not apply to components of type LineType or TransformerType but duplicate components will still not be added in this case.

Updated static test to reflect these changes and added new tests for the case where the method returns a new network (inplace=false), and the case where time-varying data is added (with_time=True_