Fix custom variable bug (#1353)
* add test test_assign_custom_variable

Tests if the optimization raises an error after adding a custom variable to the linopy model.

* fix: handle custom variable names (#1351)

Solutions of custom variables in the linopy model that can not be mapped to a network component are skipped.

Log message for user, to allow user to understand why the solution of the custom variable was not mapped.

* test: add test for warnings

---------

Co-authored-by: lkstrp <lkstrp@pm.me>